tests/wpt/tests/html/canvas/element/line-styles/2d.line.width.transformed.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.line.width.transformed</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.line.width.transformed</h1>
<p class="desc">Line stroke widths are affected by scale transformations</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:<p><img src="/images/green-100x50.png" class="output expected" id="expected" alt="">
<ul id="d"></ul>
<script>
var t = async_test("Line stroke widths are affected by scale transformations");
_addTest(function(canvas, ctx) {

  ctx.fillStyle = '#0f0';
  ctx.fillRect(0, 0, 100, 50);

  ctx.lineWidth = 4;
  // Draw a green line over a red box, to check the line is not too small
  ctx.fillStyle = '#f00';
  ctx.strokeStyle = '#0f0';
  ctx.fillRect(15, 15, 20, 20);
  ctx.save();
   ctx.scale(5, 1);
   ctx.beginPath();
   ctx.moveTo(5, 15);
   ctx.lineTo(5, 35);
   ctx.stroke();
  ctx.restore();

  // Draw a green box over a red line, to check the line is not too large
  ctx.fillStyle = '#0f0';
  ctx.strokeStyle = '#f00';
  ctx.save();
   ctx.scale(-5, 1);
   ctx.beginPath();
   ctx.moveTo(-15, 15);
   ctx.lineTo(-15, 35);
   ctx.stroke();
  ctx.restore();
  ctx.fillRect(65, 15, 20, 20);

  _assertPixel(canvas, 14,25, 0,255,0,255);
  _assertPixel(canvas, 15,25, 0,255,0,255);
  _assertPixel(canvas, 16,25, 0,255,0,255);
  _assertPixel(canvas, 25,25, 0,255,0,255);
  _assertPixel(canvas, 34,25, 0,255,0,255);
  _assertPixel(canvas, 35,25, 0,255,0,255);
  _assertPixel(canvas, 36,25, 0,255,0,255);

  _assertPixel(canvas, 64,25, 0,255,0,255);
  _assertPixel(canvas, 65,25, 0,255,0,255);
  _assertPixel(canvas, 66,25, 0,255,0,255);
  _assertPixel(canvas, 75,25, 0,255,0,255);
  _assertPixel(canvas, 84,25, 0,255,0,255);
  _assertPixel(canvas, 85,25, 0,255,0,255);
  _assertPixel(canvas, 86,25, 0,255,0,255);

});
</script>
